% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsvd.R
\name{bdsvd.ht}
\alias{bdsvd.ht}
\title{Hyperparameter Tuning for BD-SVD}
\usage{
bdsvd.ht(X, dof.lim, standardize = TRUE, anp = "2", max.iter)
}
\arguments{
\item{X}{Data matrix of dimension \eqn{n x p} with possibly \eqn{p >> n}.}

\item{dof.lim}{Interval limits for the number of non-zero components in the sparse loading (degrees of freedom).
If \eqn{S} denotes the support of \eqn{v}, then the cardinality of the support, \eqn{|S|},
corresponds to the degrees of freedom. Default is \code{dof.lim <- c(0, p-1)} which is highly recommended to check for
all levels of sparsity.}

\item{standardize}{Standardize the data to have unit variance. Default is \code{TRUE}.}

\item{anp}{Which regularization function should be used for the HBIC.
\itemize{
\item \code{"1"}: implements \eqn{a_{np} = 1} which corresponds to the BIC.
\item \code{"2"}: implements \eqn{a_{np} = 1/2 log(np)} which corresponds to the regularization used by Bauer (2025).
\item \code{"3"}: implements \eqn{a_{np} = log(log(np))}.
\item \code{"4"}: implements \eqn{a_{np} = log(log(p))} which corresponds to the regularization used by Wang et al. (2009) and Wang et al. (2013).
}}

\item{max.iter}{How many iterations should be performed for computing the sparse loading.
Default is \code{200}.}
}
\value{
\item{dof}{
The optimal number of nonzero components (degrees of freedom) according to the HBIC.
}
\item{BIC}{
The HBIC for the different numbers of nonzero components.
}
}
\description{
Finds the number of non-zero elements of the sparse loading according to the high-dimensional
Bayesian information criterion (HBIC).
}
\details{
The sparse loadings are computed using the method proposed by Shen & Huang (2008). The corresponding implementation is written in \code{Rcpp}/\code{RcppArmadillo}
for computational efficiency and is based on the \code{R} implementation by Baglama, Reichel, and Lewis in \code{\link[irlba]{ssvd}} \pkg{irlba}.
However, the implementation has been adapted to better align with the scope of the \pkg{bdsvd} package. The computation of the HBIC is outlined in Bauer (2025).
}
\examples{
#Replicate the illustrative example from Bauer (2025).


p <- 300 #Number of variables. In Bauer (2025), p = 3000
n <- 500 #Number of observations
b <- 3   #Number of blocks
design <- "c"

#Simulate data matrix X
set.seed(1)
Sigma <- bdsvd.cov.sim(p = p, b = b, design = design)
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)
colnames(X) <- seq_len(p)

ht <- bdsvd.ht(X)
plot(0:(p-1), ht$BIC[,1], xlab = "|S|", ylab = "HBIC", main = "", type = "l")
single.bdsvd(X, dof = ht$dof, standardize = FALSE)

}
\references{
\cite{Bauer, J.O. (2025). High-dimensional block diagonal covariance structure detection using singular vectors, J. Comput. Graph. Stat., 34(3), 1005–1016}

\cite{Shen, H. and Huang, J.Z. (2008). Sparse principal component analysis via regularized low rank matrix approximation, J. Multivar. Anal. 99, 1015–1034.}

\cite{Wang, H., B. Li, and C. Leng (2009). Shrinkage tuning parameter selection with a diverging number of parameters, J. R. Stat. Soc. B 71 (3), 671–683.}

\cite{Wang, L., Y. Kim, and R. Li (2013). Calibrating nonconvex penalized regression in ultra-high dimension, Ann. Stat. 41 (5), 2505–2536.}
}
\seealso{
\code{\link{bdsvd}}, \code{\link{single.bdsvd}}
}
