% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsvd.R
\docType{class}
\name{block-class}
\alias{block-class}
\title{Block}
\description{
Class used within the package to store the structure and
information about the detected blocks.
}
\section{Slots}{

\describe{
\item{\code{features}}{numeric vector that contains the the variables
corresponding to this block.}

\item{\code{block.columns}}{numeric vector that contains the indices of the
singular vectors corresponding to this block.}
}}

