% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsvd.R
\name{cov.sim}
\alias{cov.sim}
\title{Covariance matrix simulation}
\usage{
cov.sim(p = p, b = b, design = design)
}
\arguments{
\item{p}{Number of variables.}

\item{b}{Number of blocks.}

\item{design}{Simulation design "a", "b", "c", or "d".}
}
\value{
A list containing the feature names of all detected blocks.
}
\description{
This function generates covariance matrices based on the simulation studies described in Bauer (202X).
}
\examples{
#The covariance matrix for simulation design (a) is given by
Sigma <- cov.sim(p = 500, b = 500, design = "a")

}
\references{
Bauer, J.O. (202X). \emph{High-dimensional block diagonal covariance structure detection using singular vectors}\cr
}
