% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsvd.R
\name{detect.blocks}
\alias{detect.blocks}
\title{Block Detection}
\usage{
detect.blocks(V, threshold = 0)
}
\arguments{
\item{V}{Numeric matrix which either contains the loadings or is a covariance matrix.}

\item{threshold}{All absolute values of \code{V} below the threshold are set to zero.}
}
\value{
An object of class Block containing the features and columns indices corresponding to each detected block.
}
\description{
This function returns the block structure of a matrix.
}
\examples{
#In the first example, we replicate the simulation study for the ad hoc procedure
#Est_0.1 in Bauer (202X). In the second example, we manually compute the first step
#of BD-SVD, provided in the functions iterative.bdsvd() and/or bdsvd(), for constructed
#sparse loadings

#Example 1: Replicate the simulation study (a) from Bauer (202X) for the ad hoc procedure Est_0.1.
require(mvtnorm)

p <- 500 #Number of variables
n <- 125 #Number of observations
b <- 500 #Number of blocks
design <- "a"

#Simulate data matrix X
set.seed(1)
Sigma <- cov.sim(p = p, b = b, design = design)
X <- rmvnorm(n, mean=rep(0,p), sigma=Sigma)
colnames(X) <- 1:p

#Perform the ad hoc procedure
require(cvCovEst)
detect.blocks(scadEst(dat = X, lambda = 0.2), threshold = 0)

#Example 2: Manually compute the first step of BD-SVD
#for some loadings V that mirror the two blocks
#("A", "B") and c("C", "D").
require(mvtnorm)

V <- matrix(c(1,0,
              1,0,
              0,1,
              0,1), 4, 2, byrow = TRUE)

rownames(V) <- c("A", "B", "C", "D")
detected.blocks <- detect.blocks(V)

#Variables in block one with corresponding column index:
detected.blocks[[1]]@features
detected.blocks[[1]]@block.columns

#Variables in block two with corresponding column index:
detected.blocks[[2]]@features
detected.blocks[[2]]@block.columns

}
\references{
Bauer, J.O. (202X). \emph{High-dimensional block diagonal covariance structure detection using singular vectors}\cr
}
\seealso{
\link{bdsvd}, \link{iterative.bdsvd}
}
