% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcsvd.R
\name{hcsvd.cor.sim}
\alias{hcsvd.cor.sim}
\title{Correlation Matrix Simulation for HC-SVD}
\usage{
hcsvd.cor.sim(p = p, b = b, design = design)
}
\arguments{
\item{p}{Number of variables.}

\item{b}{Number of blocks.}

\item{design}{Simulation design "a" or "b".}
}
\value{
A correlation matrix according to the chosen simulation design.
}
\description{
This function generates correlation matrices based on the simulation studies described in Bauer (202Xb).
}
\examples{
#The correlation matrix for simulation design (a) is given by
#R <- hcsvd.cov.sim(p = 100, b = 5, design = "a")

}
\references{
\cite{Bauer, J.O. (202Xb). Hierarchical variable clustering using singular vectors.}
}
