% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispca.R
\name{prmats}
\alias{prmats}
\title{Principal (Sub)Matrices}
\usage{
prmats(X, block.structure, rule = "cumvar", value)
}
\arguments{
\item{X}{Data matrix of dimension \eqn{n}x\eqn{p} with possibly \eqn{p >> n}.}

\item{block.structure}{Underlying block structure. Must be a '\code{bdsvd}', '\code{blocks}', or '\code{ispca}' object. E.g., pass the result of \code{bdsvd()}, \code{single.bdsvd()},
\code{ispca()}, or \code{detect.blocks()}. An identified block structure by any other method can be supplied using \code{detect.blocks()} (see example below).}

\item{rule}{Which rule should be used to choose principal submatrices. \code{rule = "cumvar"} selects the smallest
number of principal submatrices (ordered by explained variance) whose cumulative share is at least \code{value} (a proportion in (0, 1]).
\code{rule = "enrich"} selects all principal submatrices that explain \eqn{\times}\code{value} more than they should on average (see \code{value}).}

\item{value}{Numeric parameter used by \code{rule}.
If \code{rule = "cumvar"}, \code{value} is the target cumulative proportion of explained variance (must be in \code{(0, 1]}).
Default is \code{0.8}.
If \code{rule = "enrich"}, \code{value} is the factor necessary to be selected compared to the equal-share baseline.
E.g., if a submatrix should on average explain 10\% of the total explained variance and if \code{value = 2}, this submatrix
is only selected if it explains at least 2\eqn{x}10\% = 20\% of the total explained variance. Default is \code{2}.}
}
\value{
A named list with the following components:
\describe{
\item{prmats}{List of submatrices ordered by explained variance (rule = 'cumvar') or by factor (rule = 'enrich').
Each element \code{prmats[[b]]} is a named list with:
\describe{
\item{expl.var}{
Proportion of total variance explained by block \eqn{b}.
}
\item{avg.var}{
Average variance of the variables in block \eqn{b}.
}
\item{factor}{
Enrichment factor \code{expl.var / avg.var} (see \code{value} argument of the function).
}
\item{feature.names}{
Column names (variables) that belong to block \eqn{b}.
}
\item{p.b}{
Number of variables in block \eqn{b}.
}
}
}

\item{X.pr}{
The data matrix of the kept submatrices/variables.
}
}
}
\description{
Identifies the principal (sub)matrices
}
\details{
This function selects the principal (sub)matrices as described in Bauer (2026).
}
\section{Access}{

Submatrices can be accessed with list indexing, e.g., \code{res$prmats[[1]]$feature.names} gives the variable names of the first submatrix.
}

\examples{
#Example: principal submatrices of a gene expression data set with two tissue types

if (requireNamespace("dslabs", quietly = TRUE)) {
data("tissue_gene_expression", package = "dslabs")

#We only select the two tissue types kidney (6) and liver (7)
Y <- as.numeric(tissue_gene_expression$y)
X <- scale(tissue_gene_expression$x[Y \%in\% c(6, 7), ], scale = FALSE)
Y <- Y[Y \%in\% c(6, 7)]


#First: run IS-PCA (or submit a identified block structure using bdsvd(...) or detect.blocks(...))

ispca.obj <- ispca(X = X, anp = "1")


#Second: extract the submatrices that explain at least 80\% (default value) of the total variance

res <- prmats(X, block.structure = ispca.obj)
res

#One submatix is selected which contains 236 variables (out of 500) and explains
#81.67\% of the total variance
length(res$prmats)
res$prmats[[1]]$p.b
round(res$prmats[[1]]$expl.var * 100, 2)


#Alternatively: extract the submatrices that explain five times more of the total variance
#than they should on average ('factor')

res <- prmats(X, block.structure = ispca.obj, rule = "enrich", value = 1.5)
res

#The highest 'factor' is 1.73
res <- prmats(X, block.structure = ispca.obj, rule = "enrich", value = 2)


}

}
\references{
\cite{Bauer, J.O. (2025). High-dimensional block diagonal covariance structure detection using singular vectors, J. Comput. Graph. Stat., 34(3), 1005–1016}

\cite{Bauer, J.O. (2026). Beyond regularization: inherently sparse principal component analysis. Stat. Comp.}
}
\seealso{
\code{\link{bdsvd}}, \code{\link{ispca}}
}
