% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdsvd.R
\name{result.bdsvd}
\alias{result.bdsvd}
\title{Data Matrix Structure According to the Detected Block Structure.}
\usage{
result.bdsvd(
  X,
  block.structure,
  result = c("matrix", "submatrices"),
  block.order
)
}
\arguments{
\item{X}{Data matrix of dimension \eqn{n x p} with possibly \eqn{p >> n}.}

\item{block.structure}{A user supplied block structure based on \code{iterative.bdsvd()} or \code{bdsvd()}.}

\item{result}{Should the result be the data matrix ordered according to the blocks (\code{result == "matrix"}), or
a list containing the submatrices (\code{result == "submatrices"}). Default is \code{matrix}.}

\item{block.order}{A vector that contains the order of the blocks detected by \code{iterative.bdsvd()} or \code{.bdsvd()}.
The vector must contain the index of each blocks exactly once. Default is \code{1:b} where \code{b} is the total number of blocks.}
}
\value{
Either the data matrix \code{X} with columns sorted according to the detected blocks, or a list containing the detected
submatrices.
}
\description{
Sorts the data matrix \eqn{X} according to the detected block structure \eqn{X_1 , ... , X_b} and returns
the detected submatrices individually.
}
\examples{
#Toying with the illustrative example from Bauer (202X).

require(mvtnorm)

p <- 300 #Number of variables. In Bauer (202X), p = 3000.
n <- 500 #Number of observations
b <- 3   #Number of blocks
design <- "c"

#Simulate data matrix X
set.seed(1)
Sigma <- cov.sim(p = p, b = b, design = design)
X <- rmvnorm(n, mean=rep(0,p), sigma=Sigma)
colnames(X) <- 1:p

#Compute iterative BD-SVD
bdsvd.obj <- iterative.bdsvd(X, standardize = FALSE)

#Obtain the data matrix X, sorted by the detected blocks
colnames(result.bdsvd(X, bdsvd.obj, result <- "matrix") )
colnames(result.bdsvd(X, bdsvd.obj, result <- "matrix", block.order = c(2,1,3)) )

#Obtain the detected submatrices X_1, X_2, and X_3
colnames(result.bdsvd(X, bdsvd.obj, result <- "submatrices")[[1]] )
colnames(result.bdsvd(X, bdsvd.obj, result <- "submatrices")[[2]] )
colnames(result.bdsvd(X, bdsvd.obj, result <- "submatrices")[[3]] )

}
\references{
Bauer, J.O. (202X). \emph{High-dimensional block diagonal covariance structure detection using singular vectors}\cr
}
\seealso{
\link{iterative.bdsvd}, \link{bdsvd}
}
